chore: refactor out options building for REST and GRPC
Avoids some duplicate code, lets us extend it uniformly.
